+++ angular/projects/lib/src/lib/base-widget.ts
@@ -1,5 +1,5 @@
 /**
- * gridstack-item.component.ts 10.1.2-dev
+ * gridstack-item.component.ts 10.2.0
  * Copyright (c) 2022 Alain Dumesny - see GridStack root license
  */