+++ src/utils.ts
@@ -1,5 +1,5 @@
 /**
- * utils.ts 10.1.2-dev
+ * utils.ts 10.2.0
  * Copyright (c) 2021 Alain Dumesny - see GridStack root license
  */